cicd/requirements.txt
boto3==1.14.48
docker==4.3.1
psycopg2-binary==2.8.5
python-dotenv==0.13.0
SQLAlchemy==1.3.17
testcontainers==3.0.3
